requirements.txt
google-auth-oauthlib
python-decouple
pyjwt
fastapi==0.120.0
uvicorn==0.38.0
gunicorn==23.0.0
aiohttp
aiofiles
aiosqlite